BugFix: fix missing `#include`s that show up under some build conditions
`main.cpp` needs the include for `QCommandLineParser` if the updater is
not included in the build.

`mudlet.h` needs the include for `QDir` if `Q_NO_SSL` is defined (might
happen on, particularly, Windows builds) - I did not test whether it is
a problem only if the updater is excluded or not...

I also found a BUG in Qt Creator which meant that code marked with
`#ifndef ...` was not being correctly grey-ed out in that editor.  I have
reported THAT upstream as:
https://bugreports.qt.io/browse/QTCREATORBUG-23817
and have switched the Mudlet code-base to use `#if !defined(...)` instead.

From previous recent experience with CodeFactor I was aware that it does
not like lines where there are `else`s with a miss-matching `{` or `}`
either side of them so I have put in or rearranged some `#if ... #endif`
extra stuff to provide the balancing things needed.

Unfortunately there is also an `#ifndef ...` as part of the *include-guard*
in every Mudlet class header file - but I will leave changing those files
to another occasion!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>