Enhance: add own encoders for CP437/667/737/869 for use on Win CI builds
Converts the names of the encodings (used internally) from `QString`s to
`QByteArrays` as that is what Qt uses for it's own management of them.

Take the translated Server Encoding name out of the static and thus
immutable TBuffer::csmEncodingTable and put them in a shared
`(QMap<QByteArray, QString>) mudlet::mEncodingNameMap`

Convert the preference dialogue to store the encoder required as a
QDataRole::User data item in each item rather than using the (translated)
text itself. This is more robust and simpler to handle. It also eliminared
the need for a couple of convertors to change between what is being used
internally to identify the encoders and what a user might use as the latter
are not actually used now.

As the way we provide the multi-byte decoding of Big5 is done it turns out
that we can supply both Big5-ETen (effectively what Qt provides as Big5+)
and Big5-HKCS, just by asking Qt for the correct one. So both are now
offered.

Check for each 8-bit Extended ASCII encoder on startup and do not offer
encoders that we do not provide - this should eliminate problems between
what the coder/developer has in their Qt libraries compared to what the
CI (installer builder) uses...

Four custom codecs are included here that I have written to support
the encodings listed - which is needed for Windows CI builds which uses
only Qt and NOT IUC libraries to provide encoders. These are integrated
into the above start-up checks and they are inserted into the system only
if what they can replace is not present.

Prepend an 'm ' to the displayed codec in the preferences to make it clear
that it is a Mudlet provided one (and might be slighly different {wrong!}
compared to an offical one). The `getServerEncoding()` result will give a
`M_` prefix to the same data so that they can be distinguished.

Replace the "CP874" encoding as it was erronously being called
"ISO 8859-3 (South Europe)" in the preferences and was probably clashing
with the "ISO 8859-3" real encoding. As it happens the proper CP874,
according to Wikipedia, has three codes that required TWO unicode points
to encode. This is not something that can be supported by the
`QVector<QChar>` based sysmtem that I have developed here {at least not
without some exceptional extra code) however the encoding IS very similiar
to CP1161 {a.k.a. Windows-874} and that CAN be done in this manner!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>