Revise: implement some discussed changes from Peer Reviewing
Finish off hiding the fact that we have differently named substitute
`TTextCodec`s with modified names compared to the Qt `QTextCodec`s that we
hope (but are not always rewarded with) will be present. The modified names
will only show up in the relevant Lua API functions if they are explicitly
asked for.

Removed an unwanted comment which documents a change as a result of this
PR:
"There was a mis-titled CP874 table here but it was faulty as the DB to DE
character are each encoded as TWO `QChars`  of U+0E49 to U+0E4C
respectively followed by U+FE00 - and this `QVector` based system cannot
handle them however CP1161 a.k.a. Windows-874 is the same as the
ISO 8859-11 with the Euro at 80 and 8 other changes at 85 and 91 to 97."

Also tweak some comments as pointed out in peer review.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>