Revise: improvements to the implimentation of setUserWindowTitle(...)
Move the core code to the main `TConsole` for the profile concerned rather
than doing it in the mudlet singletion. This is part of the policy I trying
to put into practice of moving functionality out of the central mudlet
class when it does not need to be done in that shared between all profiles
core.

Allow the second argument to the Lua API setWindowTitle to be ommitted for
the reset to default case as well as doing that for the empty string case.

Add detection and error reporting for the user trying to set the title on
a console that is NOT a user window (and which would not have a title bar).

Also add handling of an error condition that should not be possible in
practice but which code analysers might otherwise moan about.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>