BugFix: improve use of environmental variable WITH_SYSTEM_QTKEYCHAIN
Due to some errors in the use or not of `$` in some places things were not
working as expected. As some extra debugging messages revealed.

Also cleared up a misunderstanding about the behaviour of CMake
add_subdirectory(...) function.

After reviewing the prior attempt to do some of what this PR does I
determined that `find_package(Qt5Keychain)` should also work for the
system provided QtKeychain library - and return a Qt5Keychain_FOUND boolean
when it does. It also seems that it should be possible to determined
whether the library is located in a `qt5keychain` or `qtkeychain`
sub-directory and to report if it does not match our expectations and thus
cause a mismatch in the relevant `#include`s.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>