BugFix: get translations available again for CMake builds
A previous change to enforce the default C locale when using the Qt
`lrelease` utility so that the output was in a known (en_US) format that
we can parse with a Lua script was defective. This is because the CMake
`execute_process(...)` command uses the FIRST word as the command to
execute in a system shell and the remaining words as arguments to that
command. This meant that the actual lrelease command was not being
executed.  It is, however possible to achieve the requirement by instead
setting (on \*nix OSes) setting the LANG environmental variable temporarily
to the `C` value whilst running the command and then resetting it for the
rest of the build process.

Also, as the CMake process was running the Lua script to convert the output
from the above command from a different location the file names being used
and referred to in the output had an extra path component that was not
triggering the Lua `string:match(...)` for each line with a
language_COUNTRY code but that can be fixed with an extra `.*`. There was
also currently some extra lrelease output lines that was causing a `nil`
variable to be included in some operations - it was necessary to `or` the
variable with a `0` to get the maths to continued to work.

Whilst editing the Lua script I was able to identify all of us who had
worked on it so that that could be put into a GPL 2+ boilerplate. I also
tarted up the display of the statistics so that the numeric columns lines
up correctly, instead of the previous left alignment which is wrong for
columns of numbers!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>