Revise: load zip.so from a sub-directory on MacOS [skip appveyor]
Because the module is stored as brimworks/zip.so (to differentiate it from
the other zip.so that we are replacing for MacOs builds) we have to bundle
it in such a sub-directory. This also needs a tweak to the installer but
that is already in place for a testing run.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>