BugFix: fixup failing MacOs CI builds (#3636)
* BugFix: fixup failing MacOs CI builds

Inserts some symbolic links to account for the change of the library name
for linking from `zziplib` to `zziplib-0` on MacOs!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: add a missing double-quote from previous commit

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: put some debug output lines back in

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: fix a typo in previous commit

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: restore manual qmake LIBS addition [skip appveyor]

It shouldn't be needed but it did work in the other PR where I was hacking
around trying to solve the MacOs CI build failures and it is the only
functional difference I could see.

It is not expected that the second (the CMake one) build job will pass.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: correct error in numeric value use in file names

Also reremove the manual insertion of the libraries into the QMake LIBS
variable - as it should be done by Qt and pkg-config automagically.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: add zziblib to main project MacOs CMake libraries list

Hopefully this will fix the second (CMake) Travis CI build job.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: try a different name for the zziblib libary

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: try adding a FindZZip.cmake file to aid finding the zziplib package

The added file is Public Domain licensed from the "Ogre" (Object-oriented
Graphics Rendering Engine) project at http://www.ogre3d.org/. The file was
sourced from:
https://github.com/OGRECave/ogre/blob/master/CMake/Packages/FindZZip.cmake

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Append: add a missing file of macros for use with previously added file

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert last four commits

* Refactor: remove embedded luazip code and switch to lua-zip rock on MacOs

Reviewing the AppVeyor builds for Windows it appears that we are repeating
this use of luazip instead of lua-zip there as well. As there is some
tweaking and the compilation is done manually I have left comments about
this in the relevant CI files...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: remove a couple of remnants of luazip left behind from last commit

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert: don't change the Windows building of lua-zip right now...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert: another bit of don't change the Windows building of lua-zip now

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert: put back in a needed library for MacOs QMake builds

The link flag `-lz` is for the zlib library {https://zlib.net/}...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Testing: find macOs qmake CI build placement of zip luarock [skip appveyor]

Temporarily use a Testing PR in the mudlet/installers repo

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: correct location of temporary installer branch [skip appveyor]

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: temp installer to get lua-zip module right on MacOs [skip appveyor]

Will need a further commit to revert to standard installer repo when this
is shown to put the `zip.so` module in the right place for MacOs CI /
installer builds and the installer repo is updated to include the change.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert: restore C loading of `zip.so` for MacOs case. [skip appveyor]

The reason for this code is not clear - why is it needed on MacOs and not
other ones...?

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: load zip.so from a sub-directory on MacOS [skip appveyor]

Because the module is stored as brimworks/zip.so (to differentiate it from
the other zip.so that we are replacing for MacOs builds) we have to bundle
it in such a sub-directory. This also needs a tweak to the installer but
that is already in place for a testing run.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: take out a code fragment again [skip appveyor]

I shouldn't have put it in as it wasn't relevant and now break things.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revert back to main repo of installers

Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>