BugFix: add missing LIBS variable entry for FreeBSD (QMake only)
The introduction of a call to `sysinfo()` in the method:
`(int64_t) mudlet::getPhysicalMemoryTotal()` introduced in PR3479 requires
linking against an additional GNU compatibility (optional) library
`libsysinfo` which thus needs adding to the libraries required in the
project files.

This fix only applies to the qmake build - there are a couple of other
changes in progress in the repository which means it is premature for me
to work on that build system at the moment.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>