added translations
translations has the default translations as json file
translated has the finished translation for them as json files

if file in translated cannot be found the default file will be loaded.
The default file can be used also in crowdin as default file.
with the settings /translated/%locale_with_underscore%_%original_file_name% in the file settings