Revise: update original PR to fit in current development branch
Revised the Lua getMudletInfo() function to handle the encoder details
being `QByteArray`s instead of `QString`s. Also revised the text display
format better so that it can be copied-and-pasted somewhere else and
actually fit on the screen - this required including some code to wrap
the text at the comma separator. I think the code is good enough but a
second opinion in case there are any corner-cases would be welcomed.

Added a short explanation about the "m " prefix in the tool-tip for the
encoder selection widget in the preferences.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>