Enhance: add support for CP437 encoding (#3579)
* Enhance: add support for CP437 encoding

This encoding is often used within the font within the ROM of PC Display
cards and - with a lot of box drawing characters - is popular in some
MUDs for inline map displays.

Additionally I have also managed to include the following extra encodings
that are derived from CP437:
+ CP667 - "Mazovia" Used in MSDOS for Polish.
+ CP737 - Used in MSDOS for Greek, more popular than CP869 but is missing
  a couple of characters.
+ CP869 - Used in MSDOS for Greek, less popular than CP737 but more
  complete.

Hadn't arisen in the past but two of the new ones in the last commit are
not supported by the codecs that Qt provide in the current (5.14.2) version
of it's libraries.

* Enhance: add own encoders for CP437/667/737/869 for use on Win CI builds

Converts the names of the encodings (used internally) from `QString`s to
`QByteArrays` as that is what Qt uses for it's own management of them.

Take the translated Server Encoding name out of the static and thus
immutable TBuffer::csmEncodingTable and put them in a shared
`(QMap<QByteArray, QString>) mudlet::mEncodingNameMap`

Convert the preference dialogue to store the encoder required as a
QDataRole::User data item in each item rather than using the (translated)
text itself. This is more robust and simpler to handle. It also eliminated
the need for a couple of convertors to change between what is being used
internally to identify the encoders and what a user might use as the latter
are not actually used now.

Because of the way we provide the multi-byte decoding of Big5 is done it
turns out that we can supply both Big5-ETen (effectively what Qt provides
as Big5+) and Big5-HKCS, just by asking Qt for the correct one. So both
are now offered.

Check for each 8-bit Extended ASCII encoder on startup and do not offer
encoders that we do not provide - this should eliminate problems between
what the coder/developer has in their Qt libraries compared to what the
CI (installer builder) uses...

Four custom codecs are included here that I have written to support
the encodings listed - which is needed for Windows CI builds which uses
only Qt and NOT IUC libraries to provide encoders. These are integrated
into the above start-up checks and they are inserted into the system only
if what they can replace is not present.

Prepend an 'm ' to the displayed codec in the preferences to make it clear
that it is a Mudlet provided one (and might be slightly different {wrong!}
compared to an official one).

Replace the "CP874" encoding as it was erroneously being called
"ISO 8859-3 (South Europe)" in the preferences and was probably clashing
with the "ISO 8859-3" real encoding. As it happens the proper CP874,
according to Wikipedia, has three codes that required TWO Unicode points
to encode. This is not something that can be supported by the
`QVector<QChar>` based system that I have developed here {at least not
without some exceptional extra code) however the encoding IS very similar
to CP1161 {a.k.a. Windows-874} and that CAN be done in this manner!

Finish off hiding the fact that we have differently named substitute
`TTextCodec`s with modified names compared to the Qt `QTextCodec`s that we
hope (but are not always rewarded with) will be present. The modified names
will not show up in the relevant Lua API functions but will be accepted if they
are explicitly asked for to set the encoding.

Revise: update original PR to fit in current development branch

Revised the Lua getMudletInfo() function to handle the encoder details
being `QByteArray`s instead of `QString`s. Also revised the text display
format better so that it can be copied-and-pasted somewhere else and
actually fit on the screen - this required including some code to wrap
the text at the comma separator. I think the code is good enough but a
second opinion in case there are any corner-cases would be welcomed.

Added a short explanation about the "m " prefix in the tool-tip for the
encoder selection widget in the preferences.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>