Revise: add a library include needed just for FreeBSD
Whilst this is not essential for the main line code it helps to fix one
thing on FreeBSD and is the CMake counterpart to an existing PR
https://github.com/Mudlet/Mudlet/pull/3656 .

This being done here purely to put another commit onto this PR to try and
get the previously failed GitHub Action that is a Ubuntu CI build to
run successfully this time. The failure mechanism then was related to the
provisioning of the CI platform and NOT the code being tested.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>