Enhance: bodge support for CUF and code for (and discard) ED ESC codes
This is needed to properly display the (CP437 encoded) ANSI graphic images
that Durismud {mud.durismud.com} on port 443 (Bad number as that is
supposed to be HTTP over TLS/SSL but this Server does NOT encrypt this
connection!) or 7777 which has no encryption either - uses.

Whilst the cursor forward sequence can be emulated with the insertion of a
corresponding number of spaces the erase in display sequence cannot safely.
The default (0) option is to clear the screen forward from the current
cursor position - which is effectively a  no-operation for Mudlet; the
other two values that are codified are:
1 - clear backwards from the cursor position to the top of the screen
2 - clear the entire screen, erase the scrollback buffer and put the cursor
in the top left corner.

Neither of those behaviours are ones we can support because the screen
contents do not represent just the output from the MUD Game server and it
would be inappropriate for the Game server to be able to clear stuff like
this.

It is worth pointing out that this MUD Game Server does not attempt to
identify the Client it is connected to (it does not use the Telnet TTYPE
sub-option) so for it to be sending cursor movement/screen erasure commands
when it does not know that the client even supports them is totally broken
in my humble opinion!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>