Revert: undo user option to control login in favour of replaceable script
This puts a lot of things back to how they were and moves the existing
timed sending of the user name and password to a new `doLogin()` Lua
function which effectively does the same thing but can be replaced by
something more sophisticated if the user needs it. Alternatively they
can replace it with a dummy (empty) function with the same name and use
a pair of triggers (which I have tested) or an sysConnectionEvent event
handler.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>