Enhance: allow the use of system provided QtKeyChain packages (both QMAKE & CMAKE) (#3619)
If environmental variable `WITH_OWN_QTKEYCHAIN` is defined to (Case
insensitive) `NO` then a system package which places the relevant
files in `qt5keychain` folders will be used. If said files are not present
then builds will fail. If the variable is set to any other value or not
defined at all then Mudlet will do as it did before and download the
library from a git repository - there is a separate PR in progress at the
moment to change where from and that might need revision to combine
cleanly with this PR.

Revised to only use find_package(Qt5Keychain) for system library case, 
it seems we should not use this CMake command when we are building the
library ourselves...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>