BugFix: allow CMake builds to find LuaGlobal.lua file (Linux)
When Mudlet is built from within Qt Creator IDE.

Also fixes finding of Mudlet supplied dictionary files if system ones are
not found first and the embedded Lua Code Formatter that we ship as well.

These fixes are not going to fix things for Windows or MacOs CMake builds
if they do not work.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>