BugFix: repair faults introduced by #3705
AFAICT creating a `QSet<T>` or a `QList<T>` with a range based
initialiser using the output from a Qt API method that outputs the
other type of container must be done by making a local copy (I think)
of the output from that API call and then providing the two iterators
at the extremes of that copy. This is because such an initialisation
MUST have iterators that refer to the same instance of the container
and the container and its members must last long enough for the use to
which it is to be put. The output from something like
`QMap<T1,T2>::keys()` does not meet those requirements!

This should close #3707 .

Also fixes a copy/paste error that used the `QList<T>::begin()` for
BOTH elements of a range based initialisation in:
`(void) dlgProfilePreferences::copyMap()`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>