original geyser.add, exclude hbox/vbox/adjc overwrite
Noticed that hbox/vbox didn't work with add2 as they have their own add functions and if passed on (inheritance) was active their function were overwritten what made them not functional anymore.
(The same is true for adjustable containers)

Geyser.add is reverted to be the original (no inheritance possible for that anymore, as it's not even needed)

This commit solves this issues.