Refactor: extract out processing of Telnet DO/DONT/WILL/WONT/SB
This should make the behaviour more consistent across the various
options - and importantly IMHO make the STATUS (5) option work better
because the previous code was not setting/resetting the elements in
the `(bool)` `myOptionState[256]`, `hisOptionState[256]`,
`announcedState[256]`, `heAnnouncedState[256]` and `triedToEnable[256]`
arrays for several options that we do process - with the result I
suspect that we would report them erronously back to the Game Server
should it actually ask what the settings were.  I also suspect
that some individual `bool` members of the `Host` or `cTelnet` class
that track whether an option is in use or not may be redundant, because
we could/should instead consider the value in one of those arrays:
* `Host::mServerMXPenabled` ==> `cTelnet::hisOptionState[OPT_MXP]`
* `cTelnet::mMCCP_version_1` ==> `cTelnet::hisOptionState[OPT_COMPRESS]`
* `cTelnet::mMCCP_version_2` ==> `cTelnet::hisOptionState[OPT_COMPRESS2]`
* `cTelnet::enableATCP` ==> `cTelnet::hisOptionState[OPT_ATCP]`
* `cTelnet::enableGMCP` ==> `cTelnet::hisOptionState[OPT_GMCP]`
* `cTelnet::enableMSSP` ==> `cTelnet::hisOptionState[OPT_MSSP]`
* `cTelnet::enableMSP` ==> `cTelnet::hisOptionState[OPT_MSP]`
* `cTelnet::enableChannel102` ==> `cTelnet::hisOptionState[OPT_102]`

Refactoring this part of the `cTelnet` class has also made me aware
that we may not be fully handling the fact that sub-option negotiation
is directional - in that there is a separate option for each feature
that may be enabled or disabled in each direction independently and
separately.  Some options - like TTYPE are only activated at Mudlet
(the client's) end and some like MCCP/MCCP2 are only activated at the
Game Server's end {though MCCP**3** is Scandum's flawed - IMHO -
attempt to do the same sort of compression in the other direction -
when really it should just be negotiated by the Server **sending**
`IAC WILL MCCP2`and the client responding `IAC DO`/`DONT MCCP2` as
appropriate!} Anyhow our current `sysProtocolEnabled` and
`sysProtocolDisabled` events do not indicate which direction the
protocol has been changed for...

The reason for this PR is so that I can clean up this section of code
before I look towards trying to impliment CHARSET negotiation support.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>