Restore :new behaviour and add :new2 (#3738)
* first commit

* code form improvement, change name to add2 and inheritance

* original geyser.add, exclude hbox/vbox/adjc overwrite

Noticed that hbox/vbox didn't work with add2 as they have their own add functions and if passed on (inheritance) was active their function were overwritten what made them not functional anymore.
(The same is true for adjustable containers)

Geyser.add is reverted to be the original (no inheritance possible for that anymore, as it's not even needed)

This commit solves this issues.

* new2 and useAdd2 to decide wheter use add2 or the old add

Use new2 as constructor for a container to automatically add2 for it's elements.

if only an element should use add2 then use new2 only for that one.

If a single element in an new2 container should use the old add put useAdd2 = false in that elements constructor

(cherry picked from commit 76db16fffea757e1557331dbba3c4e56fe72dc58)

# Conflicts:
#	src/mudlet-lua/lua/geyser/GeyserAdjustableContainer.lua