Revise: make debug output less spammy for unprintable codepoints
This is a per profile option, saved with the profile, that disables the
reporting of "Unprintable" codepoints in a `TConsole` (or rather the upper
pane - i.e. a `TTextEdit` thereof). Instead, in all cases, a count of all
the problematic codepoints are printed just before the TTextEdit
is destroyed. The previous behaviour - to report every single problem as it
occurs is now enabled with a new checkbox in the Special options tab of the
preferences.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>