Revise: comment out Color Trigger Debugging Message
Whilst replaying some recording that were using Color Triggers I realised
that the debug output was getting some spam from this message which could
probably be discarded for the moment. So that it can be reactivated if
needed I have commented it out rather than removing it altogether.

This was inserted in #1571 though that replaced a previous one which sent
its output to std::cout ...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>