Refactor: remove unreachable code branch
It is pointless to code something as:
`if (condition) {...} else if (!condition) {...} else {...}`
as the `else` block is unreachable.  I suspect the original pair of
coditions might once have NOT been the compliment of each other but since
they are now the second condition and the final else block can both be
purged now.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>