Enhance: improve access to TTabBar elements
Extended the `TTabBar` class so as to return the profile name associated
with a particular index and to return the index of a tab if given the name.

Also extended the inherited `QTabBar::removeTab(int)` method so as to
properly clean up the `TTabBar` (and associated `TStyle`) when a profile is
removed, furthermore combined things so that there is now a
`QTabBar::removeTab(const QString&)` method so that a tab can be removed
without having to provide the index of it in the TabBar.

This PR is a part extracted from a larger pending work which can be dealt
with separately (handle closing of profiles whilst multi-playing) - it will
be needed (the removing tab by name part at least) for that...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>