BugFix: get select all working in TConsole/TTextEdit
It needed a `normalisedSelection()` call to transfer the values manually
assigned to `(QPoint) mDragStart` and `mDragSelectionEnd` to the `mPA` and
`mPB` ones that are used when iterating through the selection.

This then showed that it was necessary to force an update of BOTH
`TTextEdit` instances for a `TConsole` so that a selection made in one
was shown in both if they both contain selected text - which is bound to
happen if ALL the text is selected.

A further click should *clear* the selection of ALL the text and start
a new drag operation.  This is now more likely to happen, but it does not
seem totally totally correct to me yet.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>